solving 3 times recenter by useeffect because hook is mounted 3 times
